+++ sdk/securitycenter/Azure.ResourceManager.SecurityCenter/tests/Scenario/AutomationCollectionTests.cs
@@ -22,6 +22,7 @@ namespace Azure.ResourceManager.SecurityCenter.Tests
 
         public AutomationCollectionTests(bool isAsync) : base(isAsync)//, RecordedTestMode.Record)
         {
+            IgnoredHeaders.Add("Accept");
         }
 
         [SetUp]
@@ -93,7 +94,6 @@ namespace Azure.ResourceManager.SecurityCenter.Tests
         }
 
         [RecordedTest]
-
         public async Task GetAll()
         {
             string automationName = Recording.GenerateAssetName("automation");
@@ -104,35 +104,32 @@ namespace Azure.ResourceManager.SecurityCenter.Tests
         }
 
         [RecordedTest]
-        [Category("Manually")]
         public async Task Delete()
         {
+            var options = new ArmClientOptions();
+            options.SetApiVersion(SecurityAutomationResource.ResourceType, "2019-01-01-preview");
+            var playbackClient = GetArmClient(options);
             string automationName = Recording.GenerateAssetName("automation");
             var automation = await CreateSecurityAutomation(automationName);
             bool flag = await _automationCollection.ExistsAsync(automationName);
             Assert.IsTrue(flag);
 
-            await automation.DeleteAsync(WaitUntil.Completed);
-            flag = await _automationCollection.ExistsAsync(automationName);
-            Assert.IsFalse(flag);
+            if (TestEnvironment.Mode != RecordedTestMode.Playback)
+            {
+                automation = playbackClient.GetSecurityAutomationResource(automation.Id);
+                await automation.DeleteAsync(WaitUntil.Completed);
+                flag = await _automationCollection.ExistsAsync(automationName);
+                Assert.IsFalse(flag);
+            }
         }
 
-        [TestCase(null)]
-        [TestCase(false)]
         [TestCase(true)]
-        [Category("Manually")]
+        [TestCase(null)]
         public async Task AddRemoveTag(bool? useTagResource)
         {
             SetTagResourceUsage(Client, useTagResource);
             string automationName = Recording.GenerateAssetName("automation");
             var automation = await CreateSecurityAutomation(automationName);
-
-            if (useTagResource == false)
-            {
-                // if useTagResource == false, it will call automation's update method and then the uri of Automation.Data.Action is null, will throw 400 bad request
-                Assert.ThrowsAsync<RequestFailedException>(() => automation.AddTagAsync("addtagkey", "addtagvalue"));
-                return;
-            }
 
             // AddTag
             await automation.AddTagAsync("addtagkey", "addtagvalue");
